requirements.txt
2captcha-python==1.2.2
annotated-types==0.6.0
anyio==3.7.1
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.3
Brotli==1.1.0
browsermob-proxy==0.8.0
bs4==0.0.1
cairocffi==1.6.1
CairoSVG==2.7.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
cryptography==41.0.4
cssselect2==0.7.0
defusedxml==0.7.1
fastapi==0.103.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.4
kaitaistruct==0.10
lxml==4.9.3
outcome==1.2.0
Pillow==10.0.1
pyasn1==0.5.0
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
pyOpenSSL==23.2.0
pyparsing==3.1.1
PySocks==1.7.1
python-decouple==3.8
requests==2.31.0
selenium==4.14.0
selenium-wire==5.1.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
starlette==0.27.0
tinycss2==1.2.1
trio==0.22.2
trio-websocket==0.11.1
typing_extensions==4.8.0
urllib3==2.0.6
uvicorn==0.23.2
webencodings==0.5.1
wsproto==1.2.0
zstandard==0.21.0
